Remove some functionality to its own file thats shared between master and slave computers to use
